Add routing to SPA
