+++ client/src/index.jsx
@@ -1,5 +1,11 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { App } from './App';
+import { BrowserRouter } from 'react-router-dom';
+import App from './App';
 
-render(<App />, document.getElementById('root'));
+render (
+    <BrowserRouter>
+        <App />
+    </BrowserRouter>,
+    document.querySelector('#root')
+)
