+++ client/src/pages/Examples.jsx
@@ -1,0 +1,15 @@
+import React, { useState, useEffect } from 'react';
+//import axios from 'axios';
+
+import Layout from '../components/Layout';
+
+const Examples = () => (
+    <Layout children={
+        <>
+            <h1 className="text-2xl font-medium text-start py-2">Examples</h1>
+            <p>This site can be used to do the following</p>
+        </>
+    } />
+)
+
+export default Examples;
