+++ client/src/pages/index.ejs
